Update data files to include the number of particles in the first line
I made an important change in handling the list of particles.
The datafiles were not containing information about the number of particles. The code was
reading the file line by line. On each line, it was creating a Particle object and add it to
a linked list of particles. Accessing the particles were not easy since it required to traverse
a linked list.

Here is the change: now in the first line of the datafile, the number of particles is
specified so that the code can allocate a Particle[] array before reading particle
information.