Change the way particles are stored
Instead of linked lists to store the particles, the code now allocate a Particle[] array
to keep particle information. In this way, random access to particles are much more
easy and the code is more clean. To be able to do that, the datafiles now include
the number of particles in their first lines.