Add inner product calculation of dense grids
I added a functionality to calculate the sum of
pairwise multiplications of two dense grids (aka inner product).
Since the grids are assumed to be dense, the function
exploits the fact that the internal data is stored in one-dimensional
array. In this way, there is no need to use grid->getElement method.