Improve test strategy by adding micro-level tests
For testing purposes, I was writing relatively large test
functions to check various aspects of the code. The outcome
of the test was determined by the return code of the function.
For this reason, there was a one-to-one relationship between
tests and test functions.

With the addition of msm4g_test_assert function and couple
of simple global variables, I can now include many tests in just one
test function and hence reduce the time needed to complete the tests.