Add NaCl test data
The data mimics the NaCl crystal structure of alternating
charges. The sign of the charges
are decided by using the following rule

sign = (-1)^(i+j+k)

where
i,j,k = 0,...,n are the positions
and n is 1,3 or 7. The number of particles is N = (n+1)^3

n  N     box        sum of charges
-- --    ---------  --------------
1  8     2 x 2 x 2  0
3  64    4 x 4 x 4  0
7  512   4 x 4 x 4  0