Bug fix in binning strategy and improve stencil calculation
This commit contains major improvements in the code:

Improved short-range stencil calculation
----------------------------------------
By exploiting the symmetry in \kappa^l_{n}
there is an 8-fold speedup. The code is modified
to gain 48-fold speedup when the simulation box
is a cube.

A bug fix in short-range binning strategy
-----------------------------------------
The binning strategy in short-range calculation was
in inherited from the legacy vacuum boundary condition
which didn’t include wrapping.

In periodic boundary condition, the bins in the
short-range calculation should be wrapped so that
a bin can be a neighbor of a far away bin which
was not the case in vacuum boundary conditions.