Fix stencil calculation for very large relative cutoffs
An important part in stencil calculation is the
\kappa_n^l(A) (eq:32) which includes a sum over periodic
images. To reduce the computational cost, I had once
added a convergence criterion. It turned out that
this control is troublesome for especially large relative
cutoffs. With the fix, I used all the terms for which
g_l kernel is non-zero. It is slightly slower but at least
more accurate.

During the fix, I added the following potential energy components
for diagnostic purposes:

- ushort_real
- ushort_self
- ushort_csr
- ulong_self
- ulong_direct
- ulong_fourier